requirements.txt
alabaster==0.7.13
annots==0.1.2
app-context==0.9.0
apturl==0.5.2
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.0
auth==0.5.3
Babel==2.13.1
backcall==0.2.0
bcrypt==3.2.0
blessed==1.20.0
blinker==1.7.0
bootstrap-py==2.0.0
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.0
chunk==2.0.0
chunkipy==0.0.3
click==8.1.7
colorama==0.4.4
comm==0.1.4
command-not-found==0.3
config-file==0.12.0
contourpy==1.1.1
crispy-bootstrap4==2022.1
cryptography==41.0.4
cupshelpers==1.0
customtkinter==5.2.0
cycler==0.12.1
darkdetect==0.8.0
dbus-python==1.2.18
debugpy==1.8.0
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==4.2.4
django-appconf==1.0.6
django-connected==0.1.5
django-connections==0.1
django-crispy-forms==2.0
django-environ==0.11.2
django-filter==23.3
django-import-export==3.3.1
django-phonenumber-field==7.2.0
django-picklefield==3.1
django-q==1.3.9
django-reset-migrations==0.4.0
django-validators==1.0.1
djangorestframework==3.14.0
dnspython==2.4.2
docutils==0.20.1
duplicity==0.8.21
email-validator==2.1.0.post1
emoji==2.8.0
et-xmlfile==1.1.0
eventlet==0.33.3
exceptiongroup==1.2.0
executing==2.0.0
falcon==3.1.1
fasteners==0.14.1
filelock==3.12.3
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.43.1
fpdf==1.7.2
fsspec==2023.10.0
future==0.18.2
gevent==23.9.1
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.3
gunicorn==21.2.0
httplib2==0.20.2
idna==3.3
imagesize==1.4.1
importlib-metadata==4.6.4
install==1.3.5
ipykernel==6.25.2
ipython==8.16.1
itsdangerous==2.1.2
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.3
json-cfg==0.4.2
jsonfield==3.1.0
jupyter_client==8.3.1
jupyter_core==5.3.2
keyring==23.5.0
kiwisolver==1.4.5
kwonly-args==1.0.10
langdetect==1.0.9
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
logging==0.4.9.6
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.5.1
MarkupPy==1.14
MarkupSafe==2.1.5
matplotlib==3.8.0
matplotlib-inline==0.1.6
mongoengine==0.27.0
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
mysql-connector-python==8.1.0
nest-asyncio==1.5.8
netifaces==0.11.0
networkx==3.2.1
numpy==1.25.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
odfpy==1.4.1
olefile==0.46
openpyxl==3.1.2
packaging==23.2
pandas==2.0.3
paramiko==2.9.3
parso==0.8.3
pdfminer.six==20221105
pdfplumber==0.10.2
pexpect==4.8.0
pguard==0.4.0
phonenumbers==8.13.24
phonenumberslite==8.13.24
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==4.21.12
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
Pygments==2.16.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.6.0
PyMuPDF==1.23.4
PyMuPDFb==1.23.3
PyMySQL==1.1.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyPDF2==2.12.1
pypdfium2==4.20.0
pyRFC3339==1.1
pytesseract==0.3.10
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
Python-IO==0.3
python-magic==0.4.27
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.1
redis==3.5.3
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
SecretStorage==3.3.1
setuptools-scm==8.0.4
six==1.16.0
smmap==5.0.1
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
SQLAlchemy==2.0.27
sqlparse==0.4.4
stack-data==0.6.3
stanza==1.4.2
sympy==1.12
systemd-python==234
tablib==3.5.0
tomli==2.0.1
torch==2.1.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
triton==2.1.0
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.24.4
wadllib==1.3.6
waitress==2.1.2
wcwidth==0.2.8
Werkzeug==3.0.1
WTForms==3.1.2
xdg==5
xkit==0.0.0
xlrd==2.0.1
xlwt==1.3.0
zipp==1.0.0
zope.event==5.0
zope.interface==6.1
